+++ CHANGELOG.md
@@ -5,6 +5,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Changed public functions `Get-NetworkAudit` and `Send-AuditEmail` to incorporate private function `Initialize-ModuleEnv`
+- Changed public functions `Get-ADActiveUserAudit`, `Get-ADHostAudit`,`Get-ADUserLogonAudit`,`Get-ADUserPrivilegeAudit`,`Get-ADUserPrivilegeAudit`, and `Get-ADUserWildCardAudit` to incorporate private function `Install-ADModule`
+- Changed `Write-AUditLog` variable to `$LogString`. 
+
+### Fixed
+
+- Fixed error output in `Get-NetworkAudit`
+## [0.2.1] - 2023-03-11
+
 ### Added
 
 - Added Private Functions `New-GraphEmailApp` and `Send-GraphAppEmail`. Further work needed.
@@ -26,7 +37,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     $wildcardUsers = Get-ADUserWildCardAudit -WildCardIdentifier "svc" -Report -Verbose
     Merge-ADAuditZip -FilePaths $workstations, $servers, $nonWindows, $activeUsers, $privilegedUsers, $wildcardUsers -OpenDirectory
     ```
-
 
 ## [0.2.0] - 2023-02-21